requirements.txt
pillow~=11.2.0
tqdm~=4.67.1
colorama~=0.4.6